--- coa/conf/_brain.yaml
@@ -1,47 +1,0 @@
-# coa/conf/brain.yaml
-families:
-  # --- DEBIAN ---
-  debian:
-    initrd:
-      live: "update-initramfs -u -k all || update-initramfs -c -k all"
-    remaster:
-      boot_params: "boot=live components quiet splash"
-
-  # --- ARCH LINUX ---
-  archlinux:
-    initrd:
-      live:
-        command: "KVER=$(ls /lib/modules | head -n 1) && mkinitcpio -c /etc/coa-mkinitcpio.conf -k $KVER -g /boot/initrd.img"
-        setup_files:
-          "/etc/coa-mkinitcpio.conf": |
-            MODULES=(loop overlay squashfs)
-            BINARIES=()
-            FILES=()
-            # Hook completi derivati da eggs per massima compatibilità
-            HOOKS=(base udev microcode modconf kms memdisk archiso archiso_loop_mnt archiso_pxe_common archiso_pxe_nbd archiso_pxe_http archiso_pxe_nfs block filesystems keyboard keymap consolefont fsck)
-            COMPRESSION="xz"
-    remaster:
-      # Parametri kernel per l'hook archiso
-      boot_params: "archisobasedir=arch archisolabel=OA_LIVE cow_spacesize=10G"
-      # Il ponte: Arch cerca airootfs.sfs, noi gli diamo il nostro SquashFS
-      iso_links:
-        "arch/x86_64/airootfs.sfs": "../live/filesystem.squashfs"
-
-  # --- FEDORA ---
-  fedora: &fedora_family
-    initrd:
-      live:
-        command: "dracut --force --regenerate-all"
-        setup_files:
-          "/etc/dracut.conf.d/coa.conf": |
-            hostonly="no"
-            add_dracutmodules+=" dmsquash-live rootfs-block bash "
-            compress="xz"
-    remaster:
-      boot_params: "root=live:CDLABEL=OA_LIVE rd.live.image rd.live.dir=live rd.live.squashimg=filesystem.squashfs selinux=0"
-
-  # Alias per derivate Fedora
-  rhel: *fedora_family
-  centos: *fedora_family
-  rocky: *fedora_family
-  almalinux: *fedora_family
